# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Image gallery service with folder browsing, album management and tagging support.",
        "title": "Gallery API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/album/{name}": {
            "get": {
                "description": "Returns all album directories (directories containing images) recursively",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "albums"
                ],
                "summary": "List all albums under a directory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Directory path",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/core.DirNode"
                            }
                        }
                    }
                }
            }
        },
        "/api/explore/{name}": {
            "get": {
                "description": "Returns immediate contents of a directory (subdirectories, images, others)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "browse"
                ],
                "summary": "Explore a directory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Directory path",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core.SimpleDirectory"
                        }
                    }
                }
            }
        },
        "/api/image/{name}": {
            "get": {
                "description": "Returns all images recursively under the specified directory",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "List all images under a directory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Directory path",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/core.ImageNode"
                            }
                        }
                    }
                }
            }
        },
        "/api/media/{name}": {
            "get": {
                "description": "Returns all images and videos under the specified directory.\nPagination is opt-in: pass a positive `limit` to receive a MediaPage\n({items, total, offset, limit}) whose items are images and videos merged\ninto one path-sorted sequence. Without `limit` the legacy\n{images, videos} shape is returned unchanged.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "media"
                ],
                "summary": "List all media under a directory",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Directory path",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Flatten search into subdirectories (default: true)",
                        "name": "flat",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size. When present (and \u003e 0), switches the response to MediaPage",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page offset, only meaningful together with limit (default: 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/core.MediaResponse"
                        }
                    }
                }
            }
        },
        "/api/random/{name}": {
            "get": {
                "description": "Draws a batch of random media from the specified directory. Always returns a JSON array, even for count=1, and an empty array when nothing matches. Sampling is with replacement, so the same item may appear more than once.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "images"
                ],
                "summary": "Sample random media",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Directory path",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Flatten search into subdirectories (default: true)",
                        "name": "flat",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "image",
                            "video",
                            "all"
                        ],
                        "type": "string",
                        "description": "Media type to sample: image, video or all (default: image)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "How many items to sample, 1-100 (default: 1)",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/core.NodeWithParent"
                            }
                        }
                    }
                }
            }
        },
        "/api/tag": {
            "get": {
                "description": "Returns tag statistics across all images",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tags"
                ],
                "summary": "Get all tags",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/core.TagStat"
                            }
                        }
                    }
                }
            }
        },
        "/api/tree": {
            "get": {
                "description": "Returns the directory tree structure",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "browse"
                ],
                "summary": "Get folder tree",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Include leaf nodes (default: true)",
                        "name": "leaf",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "core.DirNode": {
            "type": "object",
            "properties": {
                "cover": {
                    "$ref": "#/definitions/core.ImageNode"
                },
                "directories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.Node"
                    }
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "core.ImageNode": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.TagInfo"
                    }
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "core.MediaResponse": {
            "type": "object",
            "properties": {
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.ImageNode"
                    }
                },
                "videos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.VideoNode"
                    }
                }
            }
        },
        "core.Node": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "core.NodeWithParent": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "duration_sec": {
                    "type": "number"
                },
                "height": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "parent": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.TagInfo"
                    }
                },
                "type": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
        "core.SimpleDirectory": {
            "type": "object",
            "properties": {
                "directories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.DirNode"
                    }
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.ImageNode"
                    }
                },
                "others": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.Node"
                    }
                },
                "videos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.VideoNode"
                    }
                }
            }
        },
        "core.TagInfo": {
            "type": "object",
            "properties": {
                "tag": {
                    "type": "string"
                },
                "value": {
                    "type": "integer"
                }
            }
        },
        "core.TagStat": {
            "type": "object",
            "properties": {
                "avgScore": {
                    "type": "number"
                },
                "count": {
                    "type": "integer"
                },
                "tag": {
                    "type": "string"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "core.VideoNode": {
            "type": "object",
            "properties": {
                "caption": {
                    "type": "string"
                },
                "duration_sec": {
                    "type": "number"
                },
                "height": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/core.TagInfo"
                    }
                },
                "width": {
                    "type": "integer"
                }
            }
        }
    }
}